+++ HISTORY.md
@@ -1,5 +1,9 @@
 # History
 
+## v1.0.2 September 18, 2015
+- Added `60devstips` badge for https://tips.60devs.com
+- We now use the [esnextguardian](https://github.com/bevry/esnextguardian) package, instead of just bundling it
+
 ## v1.0.1 September 16, 2015
 - Always use https